factored out common methods and attributes of TA criteria
