+++ chalice/chalicelib/criteria/aws_elb_security_groups.py
@@ -3,45 +3,17 @@ implements aws::elb::security_groups
 checkId: xSqX82fQu
 TA checks for load balancers belonging to a security group.
 """
-import json
-
-from chalicelib.criteria.criteria_default import CriteriaDefault
-from chalicelib.aws.gds_support_client import GdsSupportClient
+from chalicelib.criteria.criteria_default import TrustedAdvisorCriterion
 
 
-class ELBSecurityGroups(CriteriaDefault):
+class ELBSecurityGroups(TrustedAdvisorCriterion):
     """
     Superclass for both checks.
     """
-    active = False
-
     def __init__(self, app):
-        # attributes to overwrite in subclasses
-        self.status_string = ''
-        self.status_interval = ''
-        # attributes common in both subclasses
         self.resource_type = 'AWS::ELB::SecurityGroups'
-        self.ClientClass = GdsSupportClient
         self.check_id = 'xSqX82fQu'
-        self.language = 'en'
-        self.region = 'us-east-1'
-        self.annotation = ''
         super(ELBSecurityGroups, self).__init__(app)
-
-    def get_data(self, session, **kwargs):
-        output = self.client.describe_trusted_advisor_check_result(
-            session,
-            checkId=self.check_id,
-            language=self.language
-        )
-        self.app.log.debug(json.dumps(output))
-        return output['flaggedResources']  # will have len() == 0 if compliant or non-applicable
-
-    def translate(self, data={}):
-        return {
-            'resource_id': data.get('resourceId', ''),
-            'resource_name': data.get('metadata', ['', '', ])[1],  # trail name or empty string
-        }
 
 
 class ELBSecurityGroupsYellow(ELBSecurityGroups):
